requirements.txt
numpy==1.21.0
python-igraph==0.9.6
texttable==1.6.3
